supports: show how to DRY out constraint func
In order to ensure that the logic which calculates the constraints is
identical for the code which creates our model and the code that
pre-evaluates it, a single function must do the calculation. It is
necessary to determine which parts of that function are different
between the two call-sites. This factoring attempts to do so by allowing
the user to create a constraint evaluation function while passing in how
the model is accessed and how the output is surfaced once computed.
Then, the pre-evaluation code and model construction code can simply
change the parameters they pass.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>