supports: ensure constraint function is used identically
Although a shared function to calculate a constraint is useful, if the
user needs to pass in parameters to the function in more than one spot
and those parameters must be equal, there is still room for error. By
factoring out the parameters that must always be identical from those
that will vary by call-site, it is possible to ensure that there is
never an error due to incorrect parameter passing.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>